requirements.txt
APScheduler==3.6.3
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
dataclasses-json==0.5.6
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.1.1
paramiko==2.8.1
pycparser==2.21
PyNaCl==1.4.0
pysftp==0.2.9
python-dotenv==0.19.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
supervisor==4.2.2
tornado==6.1
tweepy==4.4.0
typing-inspect==0.7.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
